% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq_functions.R
\name{get_g2g}
\alias{get_g2g}
\title{Get human or mouse gene_ID to gene_name annotation with biomaRt}
\usage{
get_g2g(x)
}
\description{
Get human or mouse gene_ID to gene_name annotation with biomaRt
}
