% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq_functions.R
\name{get_t2g_human_v86}
\alias{get_t2g_human_v86}
\title{Get human transcript to gene annotation with EnsDb.Hsapiens.v86}
\usage{
get_t2g_human_v86()
}
\description{
Get human transcript to gene annotation with EnsDb.Hsapiens.v86
}
