% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq_functions.R
\name{get_t2g_mouse_v79}
\alias{get_t2g_mouse_v79}
\title{Get mouse transcript to gene annotation with EnsDb.Mmusculus.v79}
\usage{
get_t2g_mouse_v79()
}
\description{
Get mouse transcript to gene annotation with EnsDb.Mmusculus.v79
}
